refactor: 테스트에 필요한 부분만 인터페이스로 분리
