+++ kotlin-playground/src/main/kotlin/org/gonza/kotlinplayground/ui/OutputErrorView.kt
@@ -1,0 +1,11 @@
+package org.gonza.kotlinplayground.ui
+
+interface OutputErrorView {
+    fun printEmptyOrNullCarNameError()
+
+    fun printDuplicatedCarNameError()
+
+    fun printInvalidCarNameLengthError()
+
+    fun printTryCountError()
+}
