+++ kotlin-playground/src/main/kotlin/org/gonza/kotlinplayground/ui/OutputView.kt
@@ -1,25 +1,3 @@
 package org.gonza.kotlinplayground.ui
 
-interface OutputView {
-    fun printInputCarNameListMessage()
-
-    fun printInputTryCountMessage()
-
-    fun printCarName(carName: String)
-
-    fun printMovement(count: Int)
-
-    fun printResultMessage()
-
-    fun printWinnerMessage(winner: String)
-
-    fun printEmptyOrNullCarNameError()
-
-    fun printDuplicatedCarNameError()
-
-    fun printInvalidCarNameLengthError()
-
-    fun printTryCountError()
-
-    fun println()
-}
+interface OutputView : OutputMessageView, OutputErrorView
