+++ kotlin-playground/src/test/kotlin/org/gonza/kotlinplayground/TestOutputErrorView.kt
@@ -1,0 +1,23 @@
+package org.gonza.kotlinplayground
+
+import org.gonza.kotlinplayground.ui.OutputErrorView
+
+class TestOutputErrorView : OutputErrorView {
+    var printErrorCount = 0
+
+    override fun printEmptyOrNullCarNameError() {
+        printErrorCount++
+    }
+
+    override fun printDuplicatedCarNameError() {
+        printErrorCount++
+    }
+
+    override fun printInvalidCarNameLengthError() {
+        printErrorCount++
+    }
+
+    override fun printTryCountError() {
+        printErrorCount++
+    }
+}
